+++ docs/CHANGES.rst
@@ -14,6 +14,9 @@ Changelog
   ``api.user.revoke_roles``
   [adamcheasley]
 
+- Add ``api.user.has_permissions`` ref #172.
+  [adamcheasley]
+
 1.2.1 (2014-06-24)
 ------------------