+++ docs/CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 1.2.2dev (unreleased)
 ---------------------
 
-- fixes if a content is copied in the same folder or in a target folder, where 
+- clarified documentation for content.copy. See #185
+
+- fixes if a content is copied in the same folder or in a target folder, where
   same source id exists, the existing source(same folder) or third object
   (same id as source) gets renamed instead of the target.
   [benniboy]