+++ docs/CHANGES.rst
@@ -14,7 +14,7 @@ Changelog
   ``api.user.revoke_roles``
   [adamcheasley]
 
-- Add ``api.user.has_permissions`` ref #172.
+- Add ``api.user.has_permission`` ref #172.
   [adamcheasley]
 
 1.2.1 (2014-06-24)